requirements.txt
ccxt==4.0.74
web3==6.11.1
eth_account==0.10.0
python-dotenv==1.0.1
aiohttp==3.9.3
websockets==12.0
